requirements.txt
# L4D2-AI-Architect Requirements
# 
# Install with: pip install -r requirements.txt
#
# Note: For GPU training, install PyTorch with CUDA first:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

# ==========================================
# Core ML Dependencies
# ==========================================

# Unsloth - 2x faster fine-tuning
# Install separately: pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"

# HuggingFace ecosystem
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.25.0
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0
sentencepiece
protobuf

# ==========================================
# RL Training
# ==========================================
stable-baselines3[extra]>=2.0.0
gymnasium>=0.29.0
tensorboard>=2.14.0

# ==========================================
# Data Collection
# ==========================================
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ==========================================
# Utilities
# ==========================================
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0

# Optional: Weights & Biases logging
# wandb>=0.16.0

# ==========================================
# Development
# ==========================================
ipython>=8.0.0
jupyter>=1.0.0
black>=23.0.0
flake8>=6.0.0

# ==========================================
# Testing
# ==========================================
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-timeout>=2.0.0

# ==========================================
# Model Export
# ==========================================
# llama-cpp-python  # For GGUF export (install separately with CUDA)
safetensors>=0.4.0

# ==========================================
# Director & Bridge
# ==========================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
websockets>=12.0

# ==========================================
# OpenAI Integration
# ==========================================
openai>=1.0.0
python-dotenv>=1.0.0

# ==========================================
# Additional ML Libraries
# ==========================================
scipy>=1.10.0
scikit-learn>=1.3.0

# ==========================================
# Embeddings & RAG
# ==========================================
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0  # Use faiss-gpu for CUDA support
